size_t textlinedata

+++ edbee-lib/edbee/models/changes/textchangewithcaret.cpp
@@ -7,15 +7,15 @@
 
 namespace edbee {
 
-TextChangeWithCaret::TextChangeWithCaret(int offset, int length, const QString& text, int caret )
-    : TextChange( offset, length, text )
-    , caret_( caret )
+TextChangeWithCaret::TextChangeWithCaret(size_t offset, size_t length, const QString& text, size_t caret)
+    : TextChange(offset, length, text)
+    , caret_(caret)
 {
 }
 
 
 /// returns the caret position
-int TextChangeWithCaret::caret() const
+size_t TextChangeWithCaret::caret() const
 {
     return caret_;
 }
@@ -23,7 +23,7 @@ int TextChangeWithCaret::caret() const
 
 /// Sets the caret position
 /// @param caret the caret to set
-void TextChangeWithCaret::setCaret(int caret)
+void TextChangeWithCaret::setCaret(size_t caret)
 {
     caret_ = caret;
 }
