+++ edbee-lib/edbee/texteditorcontroller.h
@@ -100,9 +100,9 @@ signals:
 
 public slots:
 
-    void onTextChanged( edbee::TextBufferChange change, QString oldText = QString() );
-    void onSelectionChanged( edbee::TextRangeSet *oldRangeSet );
-    void onLineDataChanged( int line, int length, int newLength );
+    void onTextChanged(edbee::TextBufferChange change, QString oldText = QString());
+    void onSelectionChanged(edbee::TextRangeSet *oldRangeSet);
+    void onLineDataChanged(size_t line, size_t length, size_t newLength);
 
     void updateAfterConfigChange();
